various changes, including clrcore fixes, fixing actual MP functionality, colcache v2, user tickets for terminal:client, and some other thing
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=c660b037ab69d35c8708643978d82fbbce10ed9f

+++ client/hooks_ny/hooks/HookTestSessionFind.cpp
@@ -33,7 +33,7 @@ void FindGameStuff(int, int, int, int)
 		pSessionInfo->hostAddress.inaOnline.s_addr = hostID;
 
 		pSessionInfo->hostAddress.wPortOnline = g_netLibrary->GetHostBase() & 0xFFFF;
-		*(uint64_t*)(pSessionInfo->hostAddress.abOnline) = g_netLibrary->GetHostBase() | ((uint64_t)g_netLibrary->GetHostBase() << 32);
+		*(uint64_t*)(pSessionInfo->hostAddress.abOnline) = g_netLibrary->GetHostBase() | (g_netLibrary->GetHostBase() << 32);
 
 		sessionData.data = 1;
