+++ client/hooks_ny/live/LiveNet.cpp
@@ -12,7 +12,7 @@ DWORD __stdcall XNetGetTitleXnAddr(XNADDR * pAddr)
 	pAddr->wPortOnline = xuidBase & 0xFFFF;
 	pAddr->ina.s_addr = g_netLibrary->GetServerNetID();
 	pAddr->inaOnline.s_addr = g_netLibrary->GetServerNetID();
-	*(uint64_t*)(pAddr->abOnline) = xuidBase | ((uint64_t)xuidBase << 32);
+	*(uint64_t*)(pAddr->abOnline) = xuidBase | (xuidBase << 32);
 
 	/*if (!_stricmp(getenv("computername"), "fallarbor"))
 	{
